add get order response

+++ src/main/java/com/shop/PetProject/dtos/order/OrderDTO.java
@@ -3,6 +3,7 @@ package com.shop.PetProject.dtos.order;
 import com.shop.PetProject.models.OrderStatuses;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Builder
@@ -10,5 +11,5 @@ public record OrderDTO(Long id,
                        Set<OrderTotalDTO> orderTotals,
                        Long customerId,
                        OrderStatuses status,
-                       Integer totalPrice) {
+                       BigDecimal totalPrice) {
 }
